add ansible, packer, vagrant and virtualbox to ansible
ansible is a base package as it will be used for cluster provisioning as well.

packer, vagrant and virtualbox are dev environment related tools and will enable
us to setup or baremetal host environments

also install epel-release package first and separately. This ensures that any extra
packages installed in tasks later are found.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>