remove the temporary etcd file as part of cleanup
This ensures that when etcd is provisioning is run
again (in case of cluster-mgr workflow or netplugin installer),
the etcd will contiue to work fine.

In case when cleanup is not run (like volplugin test environment)
then this file needs to exist to handle manual etcd restarts in a
controlled environment.

Also move the the etcd tmp file location to be in tmpfs.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>