fix swarm provisioning template
swarm has a peer/replica based HA model i.e. the `--manage` flag
only needs to be run once to init the swarm cluster. The new nodes are
added only using te `--join` flag. This fix addresses this by introducing a
`swarm_bootstrap_node_addr` variable which shall be set to the address of
the node used to bootstrap the swarm cluster.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>