move role variable definitions to role defaults
We depend on ability to override some of the variables in the installer scripts
and cluster-mgr heavily so choice of defining variables matters. `defaults` are preferred
as they hold least precedence and allow ease of overriding from command line,
inventory variables etc.

For consistency we define all variables with defaults in `defaults` directory of the
respective roles

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>