remove the blanket yum upgrade task
a call to yum module with name * cause yum to try and upgrade all
packages. This causes an unnecessary upgrade of certain packages like
docker-engines, for which we want to control the version ourselves. It is
desirable to use the appropriate state=latest or a specific version info
which each yum module call than in a single blanket task.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>