systemtests: refactoring of volume primitives. See below.
The volume CRUD primitives were lousy. I moved on @mapuri's randomized
containers patches and did the same for volumes. Because of a bug in docker
(see below) we cannot create the same volume twice in a short period of time
without hitting a cache bug:

https://github.com/docker/docker/pull/24205 for more information.

This fix implements a cleaner API for managing volumes, and converts the whole
suite to use those features, including the randomized volume names.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>