apiserver,config: refactor Global configuration management (see below)
Globals were pretty ugly in how they calculated parameters. They were stateful
and it was hard to manage between the different spots in the code.

I refactored it to use a stateless design with three new functions to set
parameters in a functional way. This way instead of passing around mutable
state, code can just ask for the type of parameters it needs.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>