client: abstraction of config package part 1: see below.
Note that this is both unfinished and unintegrated. It is unit tested,
however.

This abstraction brings in a first-class home-built client that drives the
etcd client. It does this by abstracting away the details of both path
construction and client semantics.

Entities are the old notion of "config" objects that we used before.
Global, Policy are represented mostly in full now. Other objects will come
in a second patch.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>