New DB API Part #1 (See below)
This new implementation is unfinished and contains several known issues (See
below). This implementation contains a new architecture which implements a
visitor model where the visitor (the Entity) fully contains the information to
be persisted to disk, where the processor (the Client) is responsible for
processing the object into a format which is capable to be written to the k/v
store.

This abstraction is necessary for Consul support.

Known issues (will be resolved in v2):

* Hooks calls return closures which are very expensive and generated each time
  Hooks is called.
* RuntimeOptions are not capable of being watched because the keyspace is not
  sent to the object, so the volume cannot be located.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>