db: New Locking API
This implements and adds a new interface called `Lock`. Lock is presented as an
extension to Entity; one must implement a few additional methods to support it.

A few new methods were also added to acquire and free locks both with TTLs and
persistent locks.

Tests were also added for the functionality.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>